# perf(core-database-postgres): add sender&receiver indexes