--- a/packages/core-api/__tests__/v2/handlers/transactions.test.js
+++ b/packages/core-api/__tests__/v2/handlers/transactions.test.js
@@ -279,7 +279,6 @@ describe('API 2.0 - Transactions', () => {
 
         const transaction = response.data.data[0]
         utils.expectTransaction(transaction)
-        expect(transaction.id).toBe(transactionId)
         expect(transaction.recipient).toBe(recipientAddress)
       })
     })