--- a/packages/core-database-postgres/lib/migrations/20181204300000-transactions-sender-and-receiver-indexes.sql
+++ b/packages/core-database-postgres/lib/migrations/20181204300000-transactions-sender-and-receiver-indexes.sql
@@ -1,0 +1,2 @@
+CREATE INDEX IF NOT EXISTS "transactions_sender_public_key" ON ${schema~}.transactions ("sender_public_key");
+CREATE INDEX IF NOT EXISTS "transactions_recipient_id" ON ${schema~}.transactions ("recipient_id");