--- a/packages/core-database-postgres/lib/migrations/index.js
+++ b/packages/core-database-postgres/lib/migrations/index.js
@@ -8,4 +8,5 @@ module.exports = [
   loadQueryFile(__dirname, './20181129400000-add-block_id-index-to-transactions-table.sql'),
   loadQueryFile(__dirname, './20181204100000-blocks-generator_public_key-index.sql'),
   loadQueryFile(__dirname, './20181204200000-transactions-timestamp-index.sql'),
+  loadQueryFile(__dirname, './20181204300000-transactions-sender-and-receiver-indexes.sql'),
 ]